requirements.txt
Flask==2.3.3
flask-cors==4.0.0
spotipy==2.23.0
deepface==0.0.79
python-dotenv==1.0.0
pymongo==4.5.0
numpy==1.24.3
scikit-learn==1.3.0
